Move the non-limited C API of include/compile.h into include/cpython.
